neovim: migrate from packer to lazy
The newly added Scala Metals plugin was failing when I didn't have
Metals installed on a machine. Rather than force it to be installed for
every machine that I'm using Neovim on, I opted to instead only load
the Metals plugin when a Scala file is opened.

This is achieved with folke/lazy.nvim, which allows lazy-loading of
plugins on certain filetypes.